Clarifying placeholder values
For step 5 of "Trigger a remote reboot on the device using a direct method" I would suggest clarifying exactly what properties are required for the {iot hub connection string} and {deviceIdForTargetDevice} placeholder values.  For the first one it currently it says to use the string created in the previous section, which can easily be mistaken to mean the device connection string that is highlighted in the screenshot from the "Create a device identity" section. For the {deviceIdForTargetDevice} value it should just say to use "myDeviceId" since it was explicitly defined in the "Create a device identity" section.